server/requirements.txt
djangorestframework
django
djoser
django-cors-headers